Added Settlement Terms Settings
